+++ crates/node/src/cli/prometheus.rs
@@ -3,15 +3,13 @@ use clap::Args;
 /// Parameters used to config prometheus.
 #[derive(Debug, Clone, Args)]
 pub struct PrometheusParams {
-    /// Specify Prometheus exporter TCP Port.
-    #[arg(long, value_name = "PORT")]
-    pub prometheus_port: Option<u16>,
-    /// Expose Prometheus exporter on all interfaces.
-    /// Default is local.
+    /// The port used by the prometheus RPC service.
+    #[arg(long, value_name = "PORT", default_value = "9939")]
+    pub prometheus_port: u16,
+    /// Listen on all network interfaces. This usually means the prometheus server will be accessible externally.
     #[arg(long)]
     pub prometheus_external: bool,
-    /// Do not expose a Prometheus exporter endpoint.
-    /// Prometheus metric endpoint is enabled by default.
+    /// Disable the prometheus service.
     #[arg(long)]
     pub no_prometheus: bool,
 }