+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next release
 
+- fix: re-add prometheus, doc cli args, log format
 - fix(program): Changed visibility of program serializer + archived
 - fix(classes): remove the classes ordenation to allow fork rebasements
 - fix(felt): enforce Felt type as much as possible into compute_hash.rs